Canova improvements
