+++ components/profile/ProfileAchievements.tsx
@@ -2,23 +2,22 @@
 
 import { Achievement } from '@/types/authorProfile';
 import { Tooltip } from '@/components/ui/Tooltip';
-import { resolveAchievement, TierStyle } from './ProfileAchievements.utils';
+import { resolveAchievement, ResolvedAchievement, TierStyle } from './ProfileAchievements.utils';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { cn } from '@/utils/styles';
 import { faTrophyStar } from '@fortawesome/pro-light-svg-icons';
 import { faSparkles } from '@fortawesome/pro-solid-svg-icons';
 import type { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 
-type BadgeSize = 'xs' | 'sm' | 'md' | 'lg';
+type BadgeSize = 'xs' | 'sm' | 'md';
 
-const BADGE_SIZE: Record<BadgeSize, { box: string; icon: string; facetInset: string }> = {
-  xs: { box: 'h-5 w-5', icon: 'text-[8px]', facetInset: 'inset-[2px]' },
-  sm: { box: 'h-8 w-8', icon: 'text-[11px]', facetInset: 'inset-[3px]' },
-  md: { box: 'h-10 w-10', icon: 'text-[14px]', facetInset: 'inset-[3px]' },
-  lg: { box: 'h-14 w-14', icon: 'text-[20px]', facetInset: 'inset-[5px]' },
+const BADGE_SIZE: Record<BadgeSize, { box: string; iconText: string; facetInset: string }> = {
+  xs: { box: 'h-5 w-5', iconText: 'text-[8px]', facetInset: 'inset-[2px]' },
+  sm: { box: 'h-8 w-8', iconText: 'text-[11px]', facetInset: 'inset-[3px]' },
+  md: { box: 'h-10 w-10', iconText: 'text-[14px]', facetInset: 'inset-[3px]' },
 };
 
-const CRYSTAL_SHIELD_CLIP = 'polygon(50% 0%, 91% 16%, 84% 70%, 50% 100%, 16% 70%, 9% 16%)';
+const BADGE_CLIP_PATH = 'polygon(50% 0%, 91% 16%, 84% 70%, 50% 100%, 16% 70%, 9% 16%)';
 
 interface AchievementIconBadgeProps {
   icon: IconDefinition;
@@ -27,13 +26,13 @@ interface AchievementIconBadgeProps {
   className?: string;
 }
 
-function AchievementIconBadge({
+const AchievementIconBadge = ({
   icon,
   tier,
   size = 'sm',
   className,
-}: Readonly<AchievementIconBadgeProps>) {
-  const { box, icon: iconClass, facetInset } = BADGE_SIZE[size];
+}: AchievementIconBadgeProps) => {
+  const { box, iconText, facetInset } = BADGE_SIZE[size];
   return (
     <span
       className={cn(
@@ -44,72 +43,62 @@ function AchievementIconBadge({
     >
       <span
         className="absolute inset-0 shadow-sm"
-        style={{ background: tier.fill, clipPath: CRYSTAL_SHIELD_CLIP }}
+        style={{ background: tier.fill, clipPath: BADGE_CLIP_PATH }}
       />
       <span
         className={cn('absolute', facetInset)}
-        style={{ background: tier.facet, clipPath: CRYSTAL_SHIELD_CLIP }}
+        style={{ background: tier.facet, clipPath: BADGE_CLIP_PATH }}
       />
       <FontAwesomeIcon
         icon={icon}
-        className={cn('relative drop-shadow-sm', iconClass)}
-        style={{ color: tier.iconColor }}
+        className={cn('relative drop-shadow-sm', iconText)}
+        style={{ color: '#FFFFFF' }}
       />
     </span>
   );
-}
+};
 
 interface TierPillProps {
   tier: TierStyle;
   className?: string;
 }
 
-function TierPill({ tier, className }: Readonly<TierPillProps>) {
-  return (
-    <span
-      className={cn(
-        'inline-flex items-center rounded-full border px-2 py-0.5 text-[10px] font-semibold uppercase tracking-wide leading-none',
-        tier.pillBg,
-        tier.pillText,
-        tier.pillBorder,
-        className
-      )}
-    >
-      {tier.label}
-    </span>
-  );
+const TierPill = ({ tier, className }: TierPillProps) => (
+  <span
+    className={cn(
+      'inline-flex items-center rounded-full border px-2 py-0.5 text-[10px] font-semibold uppercase tracking-wide leading-none',
+      tier.pillBg,
+      tier.pillText,
+      tier.pillBorder,
+      className
+    )}
+  >
+    {tier.label}
+  </span>
+);
+
+export const AchievementsSkeleton = () => (
+  <div className="flex flex-col gap-1.5">
+    {Array.from({ length: 4 }).map((_, skeletonIndex) => (
+      <div
+        key={'achievement-skeleton-' + skeletonIndex}
+        className="flex items-center gap-2.5 py-1.5 px-1.5 rounded-lg w-full"
+      >
+        <div className="h-8 w-8 rounded-lg bg-gray-200 animate-pulse flex-shrink-0" />
+        <div className="flex-1 h-4 bg-gray-200 rounded animate-pulse" />
+        <div className="h-4 w-12 bg-gray-200 rounded-full animate-pulse" />
+      </div>
+    ))}
+  </div>
+);
+
+interface AchievementTooltipContentProps {
+  resolved: ResolvedAchievement;
 }
 
-const SKELETON_ROWS = [
-  'achievement-skeleton-1',
-  'achievement-skeleton-2',
-  'achievement-skeleton-3',
-  'achievement-skeleton-4',
-] as const;
-
-export function AchievementsSkeleton() {
-  return (
-    <div className="flex flex-col gap-1.5">
-      {SKELETON_ROWS.map((rowId) => (
-        <div key={rowId} className="flex items-center gap-2.5 py-1.5 px-1.5 rounded-lg w-full">
-          <div className="h-8 w-8 rounded-lg bg-gray-200 animate-pulse flex-shrink-0" />
-          <div className="flex-1 h-4 bg-gray-200 rounded animate-pulse" />
-          <div className="h-4 w-12 bg-gray-200 rounded-full animate-pulse" />
-        </div>
-      ))}
-    </div>
-  );
-}
-
-interface AchievementTooltipProps {
-  achievement: Achievement;
-}
-
-function AchievementTooltipContent({ achievement }: Readonly<AchievementTooltipProps>) {
-  const { meta, tier, nextTier, displayValue, targetDisplayValue, isTopTier } =
-    resolveAchievement(achievement);
-
-  const progressPct = Math.min(100, Math.max(0, achievement.pctProgress * 100));
+const AchievementTooltipContent = ({ resolved }: AchievementTooltipContentProps) => {
+  const { meta, tier, nextTier, displayValue, targetDisplayValue, progressPct, isTopTier } =
+    resolved;
 
   return (
     <div className="flex w-72 flex-col gap-3 p-1 text-left">
@@ -135,12 +124,11 @@ function AchievementTooltipContent({ achievement }: Readonly<AchievementTooltipP
           ) : (
             <span className="inline-flex items-center gap-1">
               Next:
-              <AchievementIconBadge icon={meta.icon} tier={nextTier ?? tier} size="xs" />
+              <AchievementIconBadge icon={meta.icon} tier={nextTier!} size="xs" />
             </span>
           )}
           <span className="font-medium text-gray-700">
-            {displayValue} / {targetDisplayValue}
-            {meta.unit}
+            {[displayValue, '/', targetDisplayValue, meta.unit].filter(Boolean).join(' ')}
           </span>
         </div>
         <div className="h-1.5 w-full overflow-hidden rounded-full bg-gray-100">
@@ -155,14 +143,14 @@ function AchievementTooltipContent({ achievement }: Readonly<AchievementTooltipP
       </div>
     </div>
   );
-}
+};
 
-interface AchievementsProps {
+interface ProfileAchievementsProps {
   achievements: Achievement[];
   isLoading: boolean;
 }
 
-const Achievements = ({ achievements, isLoading }: AchievementsProps) => {
+const ProfileAchievements = ({ achievements, isLoading }: ProfileAchievementsProps) => {
   if (isLoading) {
     return <AchievementsSkeleton />;
   }
@@ -183,27 +171,22 @@ const Achievements = ({ achievements, isLoading }: AchievementsProps) => {
   return (
     <div className="flex flex-col gap-1">
       {achievements.map((achievement) => {
-        const { meta, tier } = resolveAchievement(achievement);
+        const resolved = resolveAchievement(achievement);
         return (
           <Tooltip
             key={`${achievement.type}-${achievement.currentMilestoneIndex}`}
-            content={<AchievementTooltipContent achievement={achievement} />}
+            content={<AchievementTooltipContent resolved={resolved} />}
             position="bottom"
             delay={100}
             width="w-auto"
             wrapperClassName="w-full"
           >
-            <div
-              className={cn(
-                'group flex items-center gap-2.5 rounded-lg px-1.5 py-1.5 w-full cursor-default',
-                'transition-colors hover:bg-gray-50 dark:hover:bg-gray-800'
-              )}
-            >
-              <AchievementIconBadge icon={meta.icon} tier={tier} size="sm" />
+            <div className="group flex items-center gap-2.5 rounded-lg px-1.5 py-1.5 w-full cursor-default transition-colors hover:bg-gray-50 dark:hover:bg-gray-800">
+              <AchievementIconBadge icon={resolved.meta.icon} tier={resolved.tier} size="sm" />
               <span className="flex-1 min-w-0 truncate text-sm font-medium text-gray-900">
-                {meta.title}
+                {resolved.meta.title}
               </span>
-              <TierPill tier={tier} />
+              <TierPill tier={resolved.tier} />
             </div>
           </Tooltip>
         );
@@ -212,4 +195,4 @@ const Achievements = ({ achievements, isLoading }: AchievementsProps) => {
   );
 };
 
-export default Achievements;
+export default ProfileAchievements;
